extract k8s package
